requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
tensorflow==2.15.0
Pillow==9.2.0
numpy==1.23.5
torch==2.0.1
torchvision==0.15.2
gunicorn==20.1.0
Werkzeug==2.2.2  # Ensure compatibility with Flask
